ReactFlaskWithCosmosMongo/backend/requirements.txt
pymongo==3.11.0
python-dotenv==0.15.0
flask==1.1.2
